requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-py==121.0.6167.85
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.49.0
fuzzywuzzy==0.18.0
gensim==3.8.0
h11==0.14.0
haversine==2.8.0
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipykernel==6.23.1
ipython==8.12.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
llvmlite==0.41.1
lxml==4.9.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==1.3.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pymongo==4.6.2
pynndescent==0.5.11
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.13.0
six==1.16.0
smart-open==7.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
threadpoolctl==3.3.0
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
umap-learn==0.5.3
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.1
wrapt==1.16.0
wsproto==1.2.0
zipp==3.17.0
